Discord AI bot/requirements.txt
discord.py
python-dotenv
PyMuPDF
pytesseract
Pillow
nest_asyncio
google-genai
